Fix awaitables being passed into no_expect_awaiter instead of the awaiters. Make expected route in awaiter::await_resume noexcept. Fix generation in slot storage being uint16_t instead of uint32_t

+++ include/physkit/detail/arena.h
@@ -26,7 +26,7 @@ template <typename T> struct arena
     struct slot
     {
         std::optional<T> value;
-        std::uint16_t gen;
+        std::uint32_t gen;
 
         [[nodiscard]] bool available() const { return !value.has_value(); }
     };
